Fixed some unhandled events throwing exceptions.
Removed the default 'this' parameter (Ldarg_0) from HookHelper.
Made sure that the event parameters are compatible to the returned ones.
Added GWM to the SurvivalKit contributors in the credits.
Updated the event names for the new event system.
Fixed the _event variable in EventManager not being set resulting in an
Exception.
Now Mono's mscorlib is referenced in SurvivalKit.Tests.
Removed the symptom of a strange bug causing a test to fail in the
constructor of World (it wants a strong name somewhere).